+++ src/main/kotlin/sola/angel/geom2d/TParam.kt
@@ -2,7 +2,11 @@ package sola.angel.geom2d
 
 import sola.angel.nums.fuzzyEquals
 
-class TParam private constructor(val value: Double) {
+/**
+ * Parameter used for iteration which value goes from a minimum (0.0 to a maximum (1.0).
+ * Construction using factory ensures valid values which are always inside the range.
+ */
+class TParam private constructor(val value: Double) : Comparable<TParam> {
 
     /* PROPERTIES */
     val isMin = this == min
@@ -14,7 +18,16 @@ class TParam private constructor(val value: Double) {
         return value * number;
     }
 
-    /* EQUALS */
+    /* METHODS */
+    override fun compareTo(other: TParam): Int {
+        return when {
+            fuzzyEquals(value, other.value) -> return 0
+            value < other.value -> -1
+            else -> 1
+        }
+    }
+
+    /* EQUALS, HASH & TO STRING */
     override fun equals(other: Any?): Boolean {
         if (other is TParam) {
             return fuzzyEquals(value, other.value)
@@ -22,6 +35,12 @@ class TParam private constructor(val value: Double) {
 
         return false
     }
+
+    override fun hashCode(): Int {
+        return value.hashCode()
+    }
+
+    override fun toString(): String = "T: $value"
 
     /* COMPANION */
     companion object {